Refactor iterate on filenames + move location
